Let .deb package handle installing config files where possible
apt is best equiped to deal with situations where a base config
file has changed since the last package upgrade, rather than us
trying to introduce logic to handle that scenario.

+++ scripts/openli-web-provisioner-install.sh
@@ -27,21 +27,35 @@ sudo -u nobody make prefix=/usr/local install_venv
 make clean
 
 mkdir -m 700 /etc/openli-provisioner-web
-cp openli_provisioner_web/config/config.yml /etc/openli-provisioner-web/config.yml
-chmod u=rw,g=,o= /etc/openli-provisioner-web/config.yml
-cp contrib/gunicorn.py /etc/openli-provisioner-web/gunicorn.py
-cp contrib/service.env /etc/openli-provisioner-web/service.env
 
-addgroup --system openli-provisioner-web
-adduser --system --group openli-provisioner-web --shell /usr/sbin/nologin --no-create-home --home /nonexistent
+if [ ! -f /etc/openli-provisioner-web/config.yml ]; then
+    cp openli_provisioner_web/config/config.yml /etc/openli-provisioner-web/config.yml
+    chmod u=rw,g=,o= /etc/openli-provisioner-web/config.yml
+fi
+
+if [ ! -f /etc/openli-provisioner-web/gunicorn.py ]; then
+    cp contrib/gunicorn.py /etc/openli-provisioner-web/gunicorn.py
+fi
+
+if [ ! -f /etc/openli-provisioner-web/service.env ]; then
+    cp contrib/service.env /etc/openli-provisioner-web/service.env
+fi
+
+addgroup -q --system openli-provisioner-web || true
+adduser -q --system --group openli-provisioner-web --shell /usr/sbin/nologin --no-create-home --home /nonexistent || true
 chown --recursive openli-provisioner-web:openli-provisioner-web /etc/openli-provisioner-web
-cp contrib/example.service /etc/systemd/system/openli-provisioner-web.service
+
+if [ ! -f /etc/systemd/system/openli-provisioner-web.service ]; then
+    cp contrib/example.service /etc/systemd/system/openli-provisioner-web.service
+fi
 systemctl daemon-reload
 
 systemctl enable openli-provisioner-web
 systemctl start redis-server
 
-cp contrib/apache.conf /etc/apache2/sites-available/openli-provisioner-web.conf
+if [ ! -f /etc/apache2/sites-available/openli-provisioner-web.conf ]; then
+    cp contrib/apache.conf /etc/apache2/sites-available/openli-provisioner-web.conf
+fi
 
 a2enmod ssl proxy proxy_http headers
 a2ensite openli-provisioner-web
